Committed from Jim Cipar's patch on tashi-dev@i.a.o:
Currently if the cluster manager returns an unknown state for hosts or
VMs, the client throws an exception in transformState.  Confusingly
(for the user), the client program does not print a list of hosts/VMs,
or an exception message; it just prints "0".  This patch for the
client translates all unknown states to the word "Unknown" so the
client can print meaningful output even when the CM is returning bad
states.

I came across this problem when I initialized the database using a
shell script, and gave all of the hosts the state value "0".



git-svn-id: https://svn.apache.org/repos/asf/incubator/tashi/trunk@792252 13f79535-47bb-0310-9956-ffa450edef68