- adding support to query vm images via the tashi client. users can now say "tashi getImages" to be presented with a list of available VM images located in the directory as specified in the conf file.
- adding support to copy a vm image via the tashi client.
Users can now copy a VM image "tashi copyImage --src src.qcow2 --dst dst.qcow2".  This allows you to keep people out of the directory with all the VM images.

This is the initial version of these feature.  Need to keep track of who copies an image, if it is public, who can copy/use, etc.  Need to eventually keep track of this info and dump into the baseDataObject.




git-svn-id: https://svn.apache.org/repos/asf/incubator/tashi/trunk@1200047 13f79535-47bb-0310-9956-ffa450edef68